Refactor: useScheduleManagement 훅 만들기 Schedule관리 기능
